refactor: implement Link
