+++ packages/ui/lib/next/index.ts
@@ -1,0 +1,1 @@
+export * from "./components/link";
